Count sigops using generics (#9670)
* Don't skip Canopy in Regtest tests

* Support Canopy commitment bytes and adjust tests

* Don't skip Canopy in Regtest tests

* Support Canopy commitment bytes and adjust tests

* Reorder imports

* Support Canopy commitment bytes and adjust tests

* Move script tests to separate tests.rs module

* Switch zcash_script to Git dependency

* Add Sigops trait for counting transparent signature operations in txs

* Implement Sigops trait for Transaction

* Implement Sigops for UnminedTx

* Implement Sigops for CachedFfiTransaction

* Add Sigops impl for zcash_primitives Transaction

* Refactor to use Transaction::sigops instead of legacy_sigop_count

* Remove legacy_sigop_count function

* Remove unused Transaction import

Updated to use fully qualified Transaction paths to resolve ambiguity.
Removed unused Transaction import from module scope.

* Use u32 instead of u64 and rename legacy_sigop_count to sigops

* Update zcash_script to 0.3.2 and remove patch for git dependency

* Reorder imports

Standardize import ordering to improve code maintainability and reduce
merge conflicts. No functional changes introduced.

* Update zebra-script/Cargo.toml


* Clarify use of dummy values in sigops counting

* Filter coinbase inputs from script iterator in Sigops implementation

Previously, coinbase inputs yielded empty byte slices; now they are
excluded. This simplifies downstream consumers and avoids passing empty
scripts.

---------

+++ zebra-rpc/src/methods/types/get_block_template.rs
@@ -159,7 +159,7 @@ pub struct BlockTemplateResponse {
 
     /// Max legacy signature operations in the block.
     #[serde(rename = "sigoplimit")]
-    pub(crate) sigop_limit: u64,
+    pub(crate) sigop_limit: u32,
 
     /// Max block size in bytes
     #[serde(rename = "sizelimit")]
